Enemy and Entity Awake calls
The Awake() in the Enemy class overrode the Awake() in the Entity base
class, which meant that the data collected in the base class was never
populated. To correct this the Entity.Awake() as been marked as "private
protected" to allow for the child classes to call it. The Enemy.Awake()
has "base.Awake()" call to ensure that the needed data is populated and
usable.